+++ CHANGELOG.md
@@ -6,10 +6,11 @@ All notable changes to this project will be documented in this file.
 ---
 ## [1.0.0] - 2022-05-17
 ### Added
+- Implementation of ABE-GPSW/AES hybrid encryption/decryption using `abe_gpsw` v0.6.9
 - Implementation of ABE-CovCrypt/AES hybrid encryption/decryption using `cover_crypt` v1.0.2
 - Implementation of HKDF HMAC-SHA256
 ### Changed
-- Depends on `abe_gpsw` v0.6.7
+- Depends on `abe_gpsw` v0.6.9
 ### Fixed
 ### Removed
 ---